Rename extension files to align with class name
