+++ modules/extensions/src/clip/clip-extension.ts
