+++ modules/extensions/src/index.ts
@@ -1,22 +1,22 @@
-export {default as BrushingExtension} from './brushing/brushing';
-export {default as DataFilterExtension} from './data-filter/data-filter';
-export {default as Fp64Extension} from './fp64/fp64';
-export {default as PathStyleExtension} from './path-style/path-style';
-export {default as FillStyleExtension} from './fill-style/fill-style';
-export {default as ClipExtension} from './clip/clip';
-export {default as CollisionFilterExtension} from './collision-filter/collision-filter';
-export {default as MaskExtension} from './mask/mask';
-export {default as _TerrainExtension} from './terrain/terrain';
+export {default as BrushingExtension} from './brushing/brushing-extension';
+export {default as DataFilterExtension} from './data-filter/data-filter-extension';
+export {default as Fp64Extension} from './fp64/fp64-extension';
+export {default as PathStyleExtension} from './path-style/path-style-extension';
+export {default as FillStyleExtension} from './fill-style/fill-style-extension';
+export {default as ClipExtension} from './clip/clip-extension';
+export {default as CollisionFilterExtension} from './collision-filter/collision-filter-extension';
+export {default as MaskExtension} from './mask/mask-extension';
+export {default as _TerrainExtension} from './terrain/terrain-extension';
 
 // Shader module
 export {default as project64} from './fp64/project64';
 
 // Types
-export type {BrushingExtensionProps} from './brushing/brushing';
-export type {DataFilterExtensionProps} from './data-filter/data-filter';
-export type {PathStyleExtensionProps} from './path-style/path-style';
-export type {FillStyleExtensionProps} from './fill-style/fill-style';
-export type {ClipExtensionProps} from './clip/clip';
-export type {CollisionFilterExtensionProps} from './collision-filter/collision-filter';
-export type {MaskExtensionProps} from './mask/mask';
-export type {TerrainExtensionProps} from './terrain/terrain';
+export type {BrushingExtensionProps} from './brushing/brushing-extension';
+export type {DataFilterExtensionProps} from './data-filter/data-filter-extension';
+export type {PathStyleExtensionProps} from './path-style/path-style-extension';
+export type {FillStyleExtensionProps} from './fill-style/fill-style-extension';
+export type {ClipExtensionProps} from './clip/clip-extension';
+export type {CollisionFilterExtensionProps} from './collision-filter/collision-filter-extension';
+export type {MaskExtensionProps} from './mask/mask-extension';
+export type {TerrainExtensionProps} from './terrain/terrain-extension';
